Refresh in-app LTI config when we change it.
