+++ src/oasis/views_lti.py
@@ -20,23 +20,6 @@ from .lib import Users2, LTIConsumers
 L = getLogger("oasisqe")
 
 
-def update_lti_config():
-    """ pylti library wants LTI configuration set in an app variable"""
-
-    lti_consumers = LTIConsumers.all_list(active=True)
-    consumers = dict()
-    for consumer in lti_consumers:
-        consumers[consumer.consumer_key] = {
-            'secret': consumer.shared_secret
-        }
-
-    app.config['PYLTI_CONFIG'] = {
-        'consumers': consumers,
-        'roles': {
-            'admin': ['Administrator', 'urn:lti:instrole:ims/lis/Administrator'],
-            'student': ['Student', 'urn:lti:instrole:ims/lis/Student']
-        }
-    }
 
 
 def error(exception=None):
@@ -86,4 +69,4 @@ def lti_launch(lti):
         session=session)
 
 
-update_lti_config()
+LTIConsumers.update_lti_config()
